While thinking about how to test this design, I had a little bit of an epiphany. A separate file for each of the commands could make testing the actual -logic- easy, while the context will keep most of this fairlty encapsulated. There might be some nuance to this, and not every step might end up being able to be isolated out like this, but this should make a lot of the testing fairly straightforward. :) Plus, having a file for each command will make things easy to navigate!
