+++ src/command/mod.rs
@@ -2,13 +2,9 @@
 use quicli::prelude::*;
 
 mod bindgen;
-mod build;
-mod install;
 mod npm;
 
 pub use self::bindgen::wasm_bindgen_build;
-pub use self::build::cargo_build_wasm;
-pub use self::install::{cargo_install_wasm_bindgen, rustup_add_wasm_target};
 pub use self::npm::{npm_pack, npm_publish};
 
 #[derive(Debug, StructOpt)]
